Add files via upload
See the src code for this and the miniProj_I2C.ioc file for the cube setup. This is I2C and Uart code for a "MASTER" I2c. This will set up a board as Master and provide a simple serial interface. I have been testing it with a SLAVE arduino (I will post code for the slave arduino in the README).

I am not sure about code for the slave STMboard yet...